Refactor file names based on BETA10
